Split single into multiple files
Board class, manual play, mcts file, mcts_nn base for NN
